refactor: 로그인콜백페이지를 로그인폴더로 이동, 전체 페이지마다 ui폴더 제거
