+++ apps/client/src/pages/all-memo/index.ts
@@ -1,1 +1,1 @@
-export { default as AllMemoPage } from './ui/all-memo-page';
+export { default as AllMemoPage } from './all-memo-page';
