+++ apps/client/src/pages/error/error-page.css.ts
