+++ apps/client/src/pages/error/error-page.tsx
