+++ apps/client/src/pages/landing/index.ts
@@ -1,1 +1,1 @@
-export { default as LandingPage } from './ui/landing-page';
+export { default as LandingPage } from './landing-page';
