+++ apps/client/src/pages/not-found/index.ts
@@ -1,1 +1,1 @@
-export { default as NotFoundPage } from './ui/not-found-page';
+export { default as NotFoundPage } from './not-found-page';
