+++ apps/client/src/shared/router/lazy.ts
@@ -1,13 +1,9 @@
 import { lazy } from 'react';
 
-export const NewMemoPage = lazy(
-  () => import('@pages/new-memo/ui/new-memo-page'),
-);
+export const NewMemoPage = lazy(() => import('@pages/new-memo/new-memo-page'));
 export const LoginPage = lazy(() => import('@pages/login/login-page'));
-export const AllMemoPage = lazy(
-  () => import('@pages/all-memo/ui/all-memo-page'),
-);
+export const AllMemoPage = lazy(() => import('@pages/all-memo/all-memo-page'));
 export const AiResultsPage = lazy(
-  () => import('@pages/ai-results/ui/ai-results-page'),
+  () => import('@pages/ai-results/ai-results-page'),
 );
-export const LabelPage = lazy(() => import('@pages/label/ui/label-page'));
+export const LabelPage = lazy(() => import('@pages/label/label-page'));
