+++ apps/client/src/shared/router/routes/public-route.ts
@@ -1,7 +1,7 @@
 import { RouteObject } from 'react-router';
 
 import { LandingPage } from '@pages/landing';
-import LoginCallbackPage from '@pages/login-callback/login-callback-page';
+import LoginCallbackPage from '@pages/login/login-callback-page';
 
 import { LoginPage } from '../lazy';
 import { PATH } from '../path';
